+++ ui/src/components/list-view/QasListView.vue
@@ -133,7 +133,7 @@ export default {
   },
 
   created () {
-    this.fetchHandler()
+    this.mx_fetchHandler({ ...this.mx_context, url: this.url }, this.fetchList)
     this.setCurrentPage()
   },
 
@@ -141,19 +141,6 @@ export default {
     changePage () {
       const query = { ...this.$route.query, page: this.page }
       this.$router.push({ query })
-    },
-
-    fetchHandler () {
-      const hasBeforeFetch = typeof this.beforeFetch === 'function'
-
-      if (hasBeforeFetch) {
-        return this.beforeFetch({
-          payload: { ...this.mx_context, url: this.url },
-          resolve: payload => this.fetchList(payload)
-        })
-      }
-
-      this.fetchList()
     },
 
     async fetchList (filters = {}) {